+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#2371** : Change search LMDB to use MDB_NOTLS flag to not tie readers to threads.
+
 * Issue **#2371** : Fix max readers error not being shown on dashboard.
 
 * Issue **#2369** : Fix missing SQL join when reprocessing all streams matching a filter.